Change draw_seams to use the new overlay rendering system
