+++ spt/utils/ent_utils.hpp
@@ -52,13 +52,6 @@ namespace utils
 	JBData CanJB(float height);
 	bool playerEntityAvailable();
 	bool GetPunchAngleInformation(QAngle& punchAngle, QAngle& punchAngleVel);
-	void FindClosestPlane(const trace_t& tr, trace_t& out, float maxDistSqr);
-	bool TraceHit(const trace_t& tr, float maxDistSqr);
-	bool TestSeamshot(const Vector& cameraPos,
-	                  const Vector& seamPos,
-	                  const cplane_t& plane1,
-	                  const cplane_t& plane2,
-	                  QAngle& seamAngle);
 #if !defined(OE)
 	void CheckPiwSave();
 #endif
