--- spt/vgui/lines.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace vgui
-{
-	void DrawLines();
-}
